% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECOS_Phases.R
\docType{data}
\name{ECOS_Phases}
\alias{ECOS_Phases}
\title{ECOS Phases Dataset}
\format{
A data frame with 2 columns:
\describe{
  \item{Phase}{Chemical name}
  \item{Salt}{Salt name}
  ...
}
}
\source{
Runsalt v1.9 software
}
\usage{
ECOS_Phases
}
\description{
A table containing ECOS phases.
}
\keyword{datasets}
