% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_salt_AMC.R
\name{fun_salt_AMC}
\alias{fun_salt_AMC}
\title{Actual Moisture Content}
\usage{
fun_salt_AMC(
  sample_container_mass,
  initial_sample_container_mass,
  dry_sample_container_mass
)
}
\arguments{
\item{sample_container_mass}{Sample container Mass, m_{c} (g)}

\item{initial_sample_container_mass}{Initilial sample  including container Mass, m_{ics} (g)}

\item{dry_sample_container_mass}{Dry sample including container Mass, m_{dcs} (g)}
}
\value{
AMC, Actual Moisture Content (wt \%)
}
\description{
Calculate the Actual Moisture Content from sample measurement data
}
\examples{
fun_salt_AMC(8.45, 15.619, 15.198)


}
