% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_salt_calcs_old.R
\name{fun_salt_calcs_old}
\alias{fun_salt_calcs_old}
\title{Calculating salt balance from ion chromatography data (now deprecated for updated version)}
\usage{
fun_salt_calcs_old(
  salt_input,
  sample_name,
  dry_g,
  water_ml,
  chloride_ppm,
  nitrate_ppm,
  sulfate_ppm,
  sodium_ppm,
  potassium_ppm,
  calcium_ppm,
  magnesium_ppm
)
}
\arguments{
\item{salt_input}{Dataframe containing ion chromatography data}

\item{sample_name}{Sample name}

\item{dry_g}{Dry mass of sample, g}

\item{water_ml}{Water added for ion chromatography, ml}

\item{chloride_ppm}{Chloride concentration, ppm}

\item{nitrate_ppm}{Nitrate concentration, ppm}

\item{sulfate_ppm}{Sulfate concentration, ppm}

\item{sodium_ppm}{Sodium concentration, ppm}

\item{potassium_ppm}{Potassium concentration, ppm}

\item{calcium_ppm}{Calcium concentration, ppm}

\item{magnesium_ppm}{Magnesium concentration, ppm}
}
\value{
Dataframe of balanced ions for ECOS Runsalt software
}
\description{
Performs a salt balance on ion chromatography data and returns outputs for ECOS Runsalt software.

This function is now deprecated and is being replaced with a new function.
}
\examples{

salt_test |> fun_salt_calcs_old()


}
