% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_salt_HMC.R
\name{fun_salt_HMC}
\alias{fun_salt_HMC}
\title{Hygroscopic Moisture Content}
\usage{
fun_salt_HMC(
  sample_container_mass,
  dry_sample_container_mass,
  wet_sample_container_mass
)
}
\arguments{
\item{sample_container_mass}{Sample container Mass, m_{c} (g)}

\item{dry_sample_container_mass}{Dry sample including container Mass, m_{dcs} (g)}

\item{wet_sample_container_mass}{Sample mass including container at 95\%RH and 20C, m_{cs95} (g)}
}
\value{
HMC, Hygroscopic Moisture Content (wt \%)
}
\description{
Calculate the Hygroscopic Moisture Content from sample measurement data
}
\examples{
fun_salt_HMC(8.45, 15.198, 15.837)


}
