% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_salt_mileq_g.R
\name{fun_salt_mileq_g}
\alias{fun_salt_mileq_g}
\title{Milliequivalents per gram (mEq/g)}
\usage{
fun_salt_mileq_g(salt_ppm, dry_g, water_ml, mol_wts)
}
\arguments{
\item{salt_ppm}{Ion concentration in ppm}

\item{dry_g}{Sample weight, g}

\item{water_ml}{Amount of water, ml}

\item{mol_wts}{Molecular weight. Divide mol_wt by 2 if sulfate (48.03) and calcium (20.04)}
}
\value{
salt_mileq_g, milliequivalents per gram (mEq/g)
}
\description{
Milliequivalents per gram (mEq/g)
}
