% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_salt_mol.R
\name{fun_salt_mol}
\alias{fun_salt_mol}
\title{Molar concentration, mols}
\usage{
fun_salt_mol(salt_ppm, water_ml, mol_wts, salt_charges_z)
}
\arguments{
\item{salt_ppm}{Ion concentration in ppm}

\item{water_ml}{Amount of water, ml}

\item{mol_wts}{Molecular weight}

\item{salt_charges_z}{Salt charge, z}
}
\value{
salt_mol Molar concentration, mols
}
\description{
Calculate ion values as mole fraction (mols)
}
\details{
Eqn
}
\examples{
# Chloride test data
fun_salt_mol(66.824, 100, 35.453, 1)

\dontrun{
salt_test |> fun_salt_mol(chloride_ppm, mol_wts$chloride, salt_charges_z$chloride)

fun_salt_mol(14.651, 0.801, 100, mol_wts$chloride, salt_charges_z$chloride)
fun_salt_mol(17.339, 0.801, 100, mol_wts$nitrate, salt_charges_z$nitrate)
fun_salt_mol(39.923, 0.801, 100, mol_wts$sulfate, salt_charges_z$sulfate)
}


}
