% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_salt_wt.R
\name{fun_salt_wt}
\alias{fun_salt_wt}
\title{Ion content in the sample (w%)}
\usage{
fun_salt_wt(salt_ppm, dry_g, water_ml)
}
\arguments{
\item{salt_ppm}{Ion concentration in ppm}

\item{dry_g}{Sample weight, g}

\item{water_ml}{Amount of water, ml}
}
\value{
salt_wt Ion content in the sample
}
\description{
Ion content in the sample (w%)
}
