% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_wts.R
\docType{data}
\name{mol_wts}
\alias{mol_wts}
\title{Molecular weights of the species}
\format{
A data frame with 2 columns:
\describe{
  \item{chloride}{Chloride weight}
  \item{nitrate}{Nitrate weight}
  \item{sulfate}{Sulfate weight}
  \item{sodium}{Sodium weight}
  \item{potassium}{Potassium weight}
  \item{calcium}{Calcium weight}
  \item{magnesium}{Magnesium weight}
  ...
}
}
\source{
Climate
}
\usage{
mol_wts
}
\description{
A table of the molecular weights of the ion species.
}
\keyword{datasets}
