% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSaltsR_app.R
\name{runSaltsR_app}
\alias{runSaltsR_app}
\title{Run SaltsR Shiny Application}
\usage{
runSaltsR_app()
}
\arguments{
\item{example}{}
}
\value{
Shiny object
}
\description{
Run this application to test the functions and perform a salt balance
}
\examples{
\dontrun{
runSaltsR_app()
}


}
