% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salt_charges_z.R
\docType{data}
\name{salt_charges_z}
\alias{salt_charges_z}
\title{Absolute charges (z) of the ion species}
\format{
A data frame with 7 columns:
\describe{
  \item{chloride}{Chloride = 1}
  \item{nitrate}{Nitrate = 1}
  \item{sulfate}{Sulfate = 2}
  \item{sodium}{Sodium = 1}
  \item{potassium}{Potassium = 1}
  \item{calcium}{Calcium = 2}
  \item{magnesium}{Magnesium = 2}
  ...
}
}
\source{
Wikipedia
}
\usage{
salt_charges_z
}
\description{
A table of the absolute charges (z) of the ion species.
}
\keyword{datasets}
