% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salt_names.R
\docType{data}
\name{salt_names}
\alias{salt_names}
\title{Names used in the spreadsheets and cleaned versions}
\format{
A data frame with 2 columns:
\describe{
  \item{change_from}{Names used in spreadsheets}
  \item{change_to}{Tidied names used in the functions}
  ...
}
}
\source{
SaltsR
}
\usage{
salt_names
}
\description{
A table containing names of variables used in the calculations.
Contains the names used and to be changed to in tidy format.
}
\keyword{datasets}
