% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salt_test.R
\docType{data}
\name{salt_test}
\alias{salt_test}
\title{Test salt dataset}
\format{
A data frame with 10 columns:
\describe{
  \item{sample_name}{Sample name}
  \item{dry_g}{Dry g}
  \item{water_ml}{Water ml}
  \item{chloride_ppm}{Chloride ppm}
  \item{nitrate_ppm}{Nitrate ppm}
  \item{sulfate_ppm}{Sulfate ppm}
  \item{sodium_ppm}{Sodium ppm}
  \item{potassium_ppm}{Potassium ppm}
  \item{calcium_ppm}{Calcium ppm}
  \item{magnesium_ppm}{Magnesium ppm}
  ...
}
}
\source{
Climate
}
\usage{
salt_test
}
\description{
Dataset used to test the functions.
}
\keyword{datasets}
