% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaltBalance.R
\name{SaltBalance}
\alias{SaltBalance}
\title{Calculate Salt Balance}
\source{
https://zenodo.org/records/6280617#.Yic0yKvMJjE
}
\usage{
SaltBalance(m_s, V_w, conc)
}
\arguments{
\item{m_s}{Numeric. Dry Sample Mass, m_s (g)}

\item{V_w}{Numeric. Amount of water added to sample for IC, V_w (ml)}

\item{conc}{Data frame. Raw Ion Chromatography Data (IC) (mg L^(-1)) containing the following columns:
\itemize{
  \item chloride: Chloride (Cl-) Concentration
  \item nitrate: Nitrate (NO3-) Concentration
  \item sulfate: Sulfate (SO42-) Concentration
  \item sodium: Sodium (Na+) Concentration
  \item potassium: Potassium (K+) Concentration
  \item calcium: Calcium (Ca2+) Concentration
  \item magnesium: Magnesium (Mg2+) Concentration
}}
}
\value{
A list containing:
  \item{pathway}{The selected pathway (I or II)}
  \item{cat.adj}{Adjusted cations}
  \item{x_adj}{Adjusted mole fractions}
  \item{w_f}{Final weight fractions}
  \item{w_eval}{A data frame with evaluation metrics}
}
\description{
This function calculates the salt balance based on dry sample mass, water amount, and ion concentrations.
}
