% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_DataUploadServer.R
\name{shiny_dataUploadServer}
\alias{shiny_dataUploadServer}
\title{Shiny Module Server for Data Upload and Processing from Runsalt output}
\usage{
shiny_dataUploadServer(id)
}
\arguments{
\item{id}{A character string that corresponds to the ID used in the UI function for this module.}
}
\value{
A reactive expression containing the tidied data frame'
}
\description{
This function creates a Shiny module server for uploading .txt files from Runsalt
}
\examples{
\dontrun{
# In a Shiny app:
ui <- fluidPage(
  shiny_dataUploadUI("dataUpload")
 )

 server <- function(input, output, session) {
   data <- shiny_dataUploadServer("dataUpload")
   }

}



}
