% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_DataUploadUI.R
\name{shiny_dataUploadUI}
\alias{shiny_dataUploadUI}
\title{UI Module for Data Upload in Shiny}
\usage{
shiny_dataUploadUI(id)
}
\arguments{
\item{id}{A character string that defines the namespace for the module's UI elements.}
}
\value{
A `tagList` containing a `uiOutput` for file upload.
}
\description{
This function creates a Shiny UI module for uploading data files from Runsalt output.
}
\examples{
\dontrun{
ui <- fluidPage(
  shiny_dataUploadUI("dataUpload")
)
}

}
