% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyRunsalt.R
\name{tidyRunsalt}
\alias{tidyRunsalt}
\title{Read and tidy Runsalt output data}
\usage{
tidyRunsalt(file_path, X_col = "RH", Y_col = "mol", Temp_value = 20)
}
\arguments{
\item{file_path}{A character string specifying the path to the Runsalt output file.}

\item{X_col}{A character string specifying the X column variable. Defaul is "RH".}

\item{Y_col}{A character string specifying the Y column variable. Defaul is "mol".}

\item{Temp_value}{A numeric specifying the temperature (C). Default is 20.}
}
\value{
A tibble with columns:

\itemize{
  \item Salt The type of salt
  \item RH Relative Humidity
  \item mol Moles of salt
  \item filename The name of the input file
}
}
\description{
This function reads a Runsalt output file, processes it, and returns a tidy dataframe.
It separates salt and variable information, pivots the data, and adds columns for
relative humidity (RH) and moles (mol).
}
\examples{

\dontrun{
result <- tidyRunsalt("path/to/runsalt/output.txt")
head(result)
}



}
