% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_salt_mileq.R
\name{fun_salt_mileq}
\alias{fun_salt_mileq}
\title{Milliequivalents per kilogram (mEq/kg)}
\usage{
fun_salt_mileq(salt_ppm, dry_g, water_ml, mol_wts)
}
\arguments{
\item{salt_ppm}{Ion concentration in ppm}

\item{dry_g}{Sample weight, g}

\item{water_ml}{Amount of water, ml}

\item{mol_wts}{Molecular weight. Divide mol_wt by 2 if sulfate (48.03) and calcium (20.04)}
}
\value{
salt_mileq, milliequivalents per kilogram (mEq/kg)
}
\description{
Calculate initial amount of substance per ion (Eq/kg)
}
\details{
Eqn 2. Equivalents per kilogram (e_i).
}
\examples{
# Chloride test data
fun_salt_mileq(66.824, 1.128, 100, 35.453)


}
