% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_salt_mileq_g.R
\name{fun_salt_mileq_g}
\alias{fun_salt_mileq_g}
\title{Milliequivalents per gram (mEq/g)}
\usage{
fun_salt_mileq_g(salt_ppm, dry_g, water_ml, mol_wts)
}
\arguments{
\item{salt_ppm}{Ion concentration in ppm}

\item{dry_g}{Sample weight, g}

\item{water_ml}{Amount of water, ml}

\item{mol_wts}{Molecular weight. Divide mol_wt by 2 if sulfate (48.03) and calcium (20.04)}
}
\value{
salt_mileq_g, milliequivalents per gram (mEq/g)
}
\description{
Calculate initial amount of substance per ion (Eq/g). Use fun_salt_mileq for charge balance.
}
\details{
Eqn 2. Equivalents per gram (e_i_g).
}
\examples{
# Chloride test data
fun_salt_mileq_g(66.824, 1.128, 100, 35.453)

}
