% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_salt_mol.R
\name{fun_salt_mol}
\alias{fun_salt_mol}
\title{Molar concentration, mols}
\usage{
fun_salt_mol(salt_ppm, water_ml, mol_wts)
}
\arguments{
\item{salt_ppm}{Ion concentration in ppm}

\item{water_ml}{Amount of water, ml}

\item{mol_wts}{Molecular weight. Divide mol_wt by 2 if sulfate (48.03) and calcium (20.04)}
}
\value{
salt_mol Molar concentration, mols
}
\description{
Calculate ion values as mole fraction (mols)
}
\details{
Eqn
}
\examples{
# Chloride test data
fun_salt_mol(66.824, 100, 35.453)


}
