% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_wts.R
\docType{data}
\name{mol_wts}
\alias{mol_wts}
\title{Molecular weights of the species}
\format{
A data frame with 2 columns:
\describe{
  \item{chloride}{Chloride weight = 35.4527}
  \item{nitrate}{Nitrate weight = 62.0049}
  \item{sulfate}{Sulfate weight = 96.064}
  \item{sodium}{Sodium weight = 22.989768}
  \item{potassium}{Potassium weight = 39.0983}
  \item{calcium}{Calcium weight = 40.078}
  \item{magnesium}{Magnesium weight = 24.305}
  ...
}
}
\source{
Climate
}
\usage{
mol_wts
}
\description{
A table of the molecular weights of the ion species.
}
\keyword{datasets}
