% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSaltsR_app.R
\name{runSaltsR_app}
\alias{runSaltsR_app}
\title{Run SaltsR Shiny Application}
\usage{
runSaltsR_app(example)
}
\arguments{
\item{example}{}
}
\value{
Shiny object
}
\description{
Run SaltsR Shiny Application
}
\examples{
# runSaltsR_app()

}
