% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salt_charges_z.R
\docType{data}
\name{salt_charges_z}
\alias{salt_charges_z}
\title{Absolute charges (z) of the ion species}
\format{
A data frame with 7 columns:
\describe{
  \item{chloride}{Chloride charge}
  \item{nitrate}{Nitrate charge}
  \item{sulfate}{Sulfate charge}
  \item{sodium}{Sodium charge}
  \item{potassium}{Potassium charge}
  \item{calcium}{Calcium charge}
  \item{magnesium}{Magnesium charge}
  ...
}
}
\source{
Climate
}
\usage{
salt_charges_z
}
\description{
A table of the absolute charges (z) of the ion species.
}
\keyword{datasets}
